Modificados la clase User, su controller y su service y tambien modificado el teamService

+++ src/main/java/es/footleague/app/services/TeamService.java
@@ -1,10 +1,10 @@
 package es.footleague.app.services;
 
 import java.util.List;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 
 import es.footleague.app.model.Team;
 import es.footleague.app.repository.TeamRepository;
@@ -25,12 +25,16 @@ public class TeamService {
     /**
      * Busca un equipo por su ID único.
      */
-    public Team findById(Long id) {
-        return teamRepository.findById(id).orElse(null);
+    public Optional<Team> findById(Long id) {
+        return teamRepository.findById(id);
+    }
+
+    public boolean existsById(Long id) {
+        return teamRepository.existsById(id);
     }
 
     /**
-     * Guarda un equipo. Sirve tanto para crear uno nuevo 
+     * Guarda un equipo. Sirve tanto para crear uno nuevo
      * como para actualizar uno existente.
      */
     public Team save(Team team) { // Cambiado de void a Team
@@ -41,6 +45,8 @@ public class TeamService {
      * Elimina un equipo de la base de datos por su ID.
      */
     public void deleteById(Long id) {
-        teamRepository.deleteById(id);
+        if (teamRepository.existsById(id)) {
+            teamRepository.deleteById(id);
+        }
     }
 }
